requirements.txt
flask>=2.3.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4
matplotlib==3.8.2
pillow>=10.0.0
joblib>=1.3.2
requests>=2.31.0
gunicorn>=21.2.0
